feat: refactor TestConfig class to enable out of tree kernel testing
